requirements.txt

pyspark
kafka-python
psycopg2-binary
PyYAML
pytest
